+++ test/ToneJSMidiOut.test.ts
@@ -5,6 +5,13 @@ import ToneJSMidiOut from '../src/ToneJSMidiOut';
 import { parsePitch } from '../src/utils';
 
 
+class ToneTest {
+    now(): string {
+        return 'now';
+    }
+}
+
+
 class TestBaseSynth {
     name: string;
     
@@ -21,11 +28,11 @@ class TestMonoSynth extends TestBaseSynth {
     }
 
     triggerAttack(frequency: number, time: any, velocity: number = 1) {
-        this.messages.push(`triggerAttack(${frequency.toFixed(0)}, null, ${velocity.toFixed(2)})`);
+        this.messages.push(`triggerAttack(${frequency.toFixed(0)}, ${time}, ${velocity.toFixed(2)})`);
     }
 
     triggerRelease(time: any) {
-        this.messages.push(`triggerRelease(null)`);
+        this.messages.push(`triggerRelease(${time})`);
     }
 }
 
@@ -35,11 +42,11 @@ class TestNoiseSynth extends TestBaseSynth {
     }
 
     triggerAttack(time: any, velocity: number = 1) {
-        this.messages.push(`triggerAttack(null, ${velocity.toFixed(2)})`);
+        this.messages.push(`triggerAttack(${time}, ${velocity.toFixed(2)})`);
     }
 
     triggerRelease(time: any) {
-        this.messages.push(`triggerRelease(null)`);
+        this.messages.push(`triggerRelease(${time})`);
     }
 }
 
@@ -49,11 +56,11 @@ class TestPolySynth extends TestBaseSynth {
     }
 
     triggerAttack(notes: number, time: any, velocity: number = 1) {
-        this.messages.push(`triggerAttack(${notes.toFixed(0)}, null, ${velocity.toFixed(2)})`);
+        this.messages.push(`triggerAttack(${notes.toFixed(0)}, ${time}, ${velocity.toFixed(2)})`);
     }
 
     triggerRelease(notes: number, time: any) {
-        this.messages.push(`triggerRelease(${notes.toFixed(0)}, null)`);
+        this.messages.push(`triggerRelease(${notes.toFixed(0)}, ${time})`);
     }
 }
 
@@ -62,7 +69,7 @@ class TestPolySynth extends TestBaseSynth {
 
 @suite class ToneJSMidiOutTests {
     @test 'setChannel populates default values'() {
-        const midiOut = new ToneJSMidiOut();
+        const midiOut = new ToneJSMidiOut(new ToneTest());
         const target = new TestMonoSynth();
         midiOut.setChannel(0, target);
         expect(midiOut.channels[0].target).to.equal(target);
@@ -70,13 +77,13 @@ class TestPolySynth extends TestBaseSynth {
     }
 
     @test 'setChannel errors if channel not in range 0 to 15'() {
-        const midiOut = new ToneJSMidiOut();
+        const midiOut = new ToneJSMidiOut(new ToneTest());
         expect(() => midiOut.setChannel(-1, new TestMonoSynth())).to.throw();
         expect(() => midiOut.setChannel(16, new TestMonoSynth())).to.throw();
     }
 
     @test 'If setChannel target is null, then entire channel object is nulled out'() {
-        const midiOut = new ToneJSMidiOut();
+        const midiOut = new ToneJSMidiOut(new ToneTest());
         midiOut.setChannel(0, new TestMonoSynth());
         expect(midiOut.channels[0]).to.not.be.null;
         midiOut.setChannel(0, null);
@@ -84,7 +91,7 @@ class TestPolySynth extends TestBaseSynth {
     }
 
     @test 'addNote causes triggerAttack to be called on corresponding channel'() {
-        const midiOut = new ToneJSMidiOut();
+        const midiOut = new ToneJSMidiOut(new ToneTest());
         const target0 = new TestMonoSynth();
         const target1 = new TestMonoSynth();
         midiOut.setChannel(0, target0);
@@ -92,11 +99,11 @@ class TestPolySynth extends TestBaseSynth {
         midiOut.addNote(new Note(60, 80, 1));
         expect(target0.messages.length).to.equal(0);
         expect(target1.messages.length).to.equal(1);
-        expect(target1.messages[0]).to.equal('triggerAttack(262, null, 0.63)');
+        expect(target1.messages[0]).to.equal('triggerAttack(262, now, 0.63)');
     }
 
     @test 'addNote does nothing if nothing set for note channel'() {
-        const midiOut = new ToneJSMidiOut();
+        const midiOut = new ToneJSMidiOut(new ToneTest());
         const target = new TestMonoSynth();
         midiOut.setChannel(0, target);
         midiOut.addNote(new Note(40, 80, 2));
@@ -105,7 +112,7 @@ class TestPolySynth extends TestBaseSynth {
     }
 
     @test 'New channel automatically gets configured to be polyphonic based on target name'() {
-        const midiOut = new ToneJSMidiOut();
+        const midiOut = new ToneJSMidiOut(new ToneTest());
         midiOut.setChannel(0, new TestMonoSynth('Synth'));
         expect(midiOut.channels[0].isPolyphonic).to.be.false;
         
@@ -120,38 +127,38 @@ class TestPolySynth extends TestBaseSynth {
     }
 
     @test 'update method finding off note will cause triggerRelease on the corresponding channel'() {
-        const midiOut = new ToneJSMidiOut();
+        const midiOut = new ToneJSMidiOut(new ToneTest());
         const target = new TestMonoSynth('Synth');
         midiOut.setChannel(0, target);
         midiOut.addNote(new Note('A4', 127, 0)).on = false;
         midiOut.update(1);
         expect(target.messages.length).to.equal(2);
-        expect(target.messages[0]).to.equal('triggerAttack(440, null, 1.00)');
-        expect(target.messages[1]).to.equal('triggerRelease(null)');
+        expect(target.messages[0]).to.equal('triggerAttack(440, now, 1.00)');
+        expect(target.messages[1]).to.equal('triggerRelease(now)');
     }
 
     @test 'triggerAttack gets called slightly differently for NoiseSynth'() {
-        const midiOut = new ToneJSMidiOut();
+        const midiOut = new ToneJSMidiOut(new ToneTest());
         const target = new TestNoiseSynth();
         midiOut.setChannel(0, target);
         midiOut.addNote(new Note('A4', 127, 0));
         expect(target.messages.length).to.equal(1);
-        expect(target.messages[0]).to.equal('triggerAttack(null, 1.00)');
+        expect(target.messages[0]).to.equal('triggerAttack(now, 1.00)');
     }
 
     @test 'triggerRelease gets called differently for polyphonic synths'() {
-        const midiOut = new ToneJSMidiOut();
+        const midiOut = new ToneJSMidiOut(new ToneTest());
         const target = new TestPolySynth('PolySynth');
         midiOut.setChannel(0, target);
         const note = midiOut.addNote(new Note('A4', 127, 0));
         note.on = false;
         midiOut.update(1);
         expect(target.messages.length).to.equal(2);
-        expect(target.messages[1]).to.equal('triggerRelease(440, null)');
+        expect(target.messages[1]).to.equal('triggerRelease(440, now)');
     }
 
     @test 'stopNotes causes all notes to be stopped on next update'() {
-        const midiOut = new ToneJSMidiOut();
+        const midiOut = new ToneJSMidiOut(new ToneTest());
         const target = new TestPolySynth('PolySynth');
         midiOut.setChannel(0, target);
         midiOut.addNote(new Note('A4', 127, 0));
@@ -160,12 +167,12 @@ class TestPolySynth extends TestBaseSynth {
         expect(target.messages.length).to.equal(2);
         midiOut.update(1);
         expect(target.messages.length).to.equal(4);
-        expect(target.messages[2]).to.equal('triggerRelease(440, null)');
-        expect(target.messages[3]).to.equal('triggerRelease(494, null)');
+        expect(target.messages[2]).to.equal('triggerRelease(440, now)');
+        expect(target.messages[3]).to.equal('triggerRelease(494, now)');
     }
 
     @test 'sendMessage for NoteOnMessage creates new note if for valid channel'() {
-        const midiOut = new ToneJSMidiOut();
+        const midiOut = new ToneJSMidiOut(new ToneTest());
         const target = new TestPolySynth('PolySynth');
         midiOut.setChannel(0, target);
         midiOut.sendMessage(new NoteOnMessage('A4', 127, 0));
@@ -173,11 +180,11 @@ class TestPolySynth extends TestBaseSynth {
         expect(midiOut.notes.length).to.equal(1);
         expect(midiOut.notes[0].pitch).to.equal(parsePitch('A4'));
         expect(target.messages.length).to.equal(1);
-        expect(target.messages[0]).to.equal('triggerAttack(440, null, 1.00)');
+        expect(target.messages[0]).to.equal('triggerAttack(440, now, 1.00)');
     }
 
     @test 'sendMessage for NoteOffMessage stops note if for valid channel'() {
-        const midiOut = new ToneJSMidiOut();
+        const midiOut = new ToneJSMidiOut(new ToneTest());
         const target = new TestPolySynth('Sampler');
         midiOut.setChannel(0, target);
         midiOut.addNote(new Note('A4', 127, 0));
@@ -185,11 +192,11 @@ class TestPolySynth extends TestBaseSynth {
         midiOut.update(1);
         expect(midiOut.notes.length).to.equal(0);
         expect(target.messages.length).to.equal(2);
-        expect(target.messages[1]).to.equal('triggerRelease(440, null)');
+        expect(target.messages[1]).to.equal('triggerRelease(440, now)');
     }
 
     @test 'sendMessage for ControlChangeMessage triggers channel onControlChange'() {
-        const midiOut = new ToneJSMidiOut();
+        const midiOut = new ToneJSMidiOut(new ToneTest());
         const target = new TestNoiseSynth();
         midiOut.setChannel(0, target);
         let ccMessage = null;
@@ -203,7 +210,7 @@ class TestPolySynth extends TestBaseSynth {
     }
 
     @test 'sendMessage for ControlChangeMessage updates controlValues array'() {
-        const midiOut = new ToneJSMidiOut();
+        const midiOut = new ToneJSMidiOut(new ToneTest());
         const target = new TestNoiseSynth();
         midiOut.setChannel(0, target);
         midiOut.sendMessage(new ControlChangeMessage(16, 90, 0));
@@ -211,7 +218,7 @@ class TestPolySynth extends TestBaseSynth {
     }
 
     @test 'sendMessage for ControlChangeMessage validates parameter values'() {
-        const midiOut = new ToneJSMidiOut();
+        const midiOut = new ToneJSMidiOut(new ToneTest());
         const target = new TestNoiseSynth();
         midiOut.setChannel(0, target);
         expect(() => midiOut.sendMessage(new ControlChangeMessage(-1, 0, 0))).to.throw();
@@ -221,7 +228,7 @@ class TestPolySynth extends TestBaseSynth {
     }
 
     @test 'sendRawData can process note-on messages'() {
-        const midiOut = new ToneJSMidiOut();
+        const midiOut = new ToneJSMidiOut(new ToneTest());
         const target = new TestPolySynth('Sampler');
         midiOut.setChannel(0, target);
         midiOut.sendRawData([0x90 + 0, parsePitch('A4'), 127]);
@@ -229,11 +236,11 @@ class TestPolySynth extends TestBaseSynth {
         expect(midiOut.notes[0].pitch).to.equal(parsePitch('A4'));
         expect(midiOut.notes[0].velocity).to.equal(127);
         expect(target.messages.length).to.equal(1);
-        expect(target.messages[0]).to.equal('triggerAttack(440, null, 1.00)');
+        expect(target.messages[0]).to.equal('triggerAttack(440, now, 1.00)');
     }
 
     @test 'sendRawData can process note-off messages'() {
-        const midiOut = new ToneJSMidiOut();
+        const midiOut = new ToneJSMidiOut(new ToneTest());
         const target = new TestPolySynth('Sampler');
         midiOut.setChannel(0, target);
         midiOut.addNote(new Note('A4', 127, 0));
@@ -241,14 +248,30 @@ class TestPolySynth extends TestBaseSynth {
         midiOut.update(1);
         expect(midiOut.notes.length).to.equal(0);
         expect(target.messages.length).to.equal(2);
-        expect(target.messages[1]).to.equal('triggerRelease(440, null)');
+        expect(target.messages[1]).to.equal('triggerRelease(440, now)');
     }
 
     @test 'sendRawData can process control change messages'() {
-        const midiOut = new ToneJSMidiOut();
+        const midiOut = new ToneJSMidiOut(new ToneTest());
         const target = new TestNoiseSynth();
         midiOut.setChannel(0, target);
         midiOut.sendRawData([0xB0 + 0, 60, 70]);
         expect(midiOut.channels[0].controlValues[60]).to.equal(70);
+    }
+
+    @test 'constructor takes tonejs reference'() {
+        const tonejs = new ToneTest();
+        const midiOut = new ToneJSMidiOut(tonejs);
+        expect(midiOut.toneJS).to.equal(tonejs);
+    }
+
+    @test 'constructor throws error if toneJS parameter is falsy'() {
+        expect(() => new ToneJSMidiOut(null)).to.throw();
+    }
+
+    @test 'ToneJSMidiOutChannels store reference to their parent'() {
+        const midiOut = new ToneJSMidiOut(new ToneTest());
+        midiOut.setChannel(0, new TestNoiseSynth());
+        expect(midiOut.channels[0].parent).to.equal(midiOut);
     }
 }